+++ services/sven-eidolon/src/types.ts
@@ -118,6 +118,11 @@ export type EidolonBuildingKind =
   | 'lb_orchestrator'
   | 'cdn_proxy'
   | 'rate_controller'
+  | 'log_streamer'
+  | 'metrics_hub'
+  | 'event_correlator'
+  | 'trace_collector'
+  | 'dashboard_builder'
   | 'log_rotator'
   | 'ip_gatekeeper'
   | 'webhook_relay'
@@ -777,6 +782,26 @@ export type EidolonEventKind =
   | 'regmgr.repo_created' | 'regmgr.tag_pushed' | 'regmgr.retention_applied' | 'regmgr.vuln_found'
   | 'imgbld.build_started' | 'imgbld.build_completed' | 'imgbld.image_pushed' | 'imgbld.build_failed'
   | 'artstore.artifact_uploaded' | 'artstore.artifact_downloaded' | 'artstore.retention_applied' | 'artstore.access_audited' | 'plrun.pipeline_triggered' | 'plrun.stage_completed' | 'plrun.run_passed' | 'plrun.run_failed' | 'torch.suite_executed' | 'torch.failure_analyzed' | 'torch.coverage_tracked' | 'torch.report_exported' | 'dpmgr.deploy_initiated' | 'dpmgr.health_verified' | 'dpmgr.deploy_promoted' | 'dpmgr.instances_drained' | 'rbctl.rollback_initiated' | 'rbctl.snapshot_restored' | 'rbctl.rollback_verified' | 'rbctl.report_exported' | 'relgk.candidate_evaluated' | 'relgk.gate_checked' | 'relgk.release_promoted' | 'relgk.candidate_rejected' | 'mexp.metric_recorded' | 'mexp.alert_triggered' | 'mexp.dashboard_exported' | 'mexp.alerts_listed' | 'lship.pipeline_created' | 'lship.logs_shipped' | 'lship.health_checked' | 'lship.config_exported' | 'almgr.rule_created' | 'almgr.alert_fired' | 'almgr.incident_acknowledged' | 'almgr.incident_resolved' | 'incrs.incident_opened' | 'incrs.diagnosis_completed' | 'incrs.remediation_executed' | 'incrs.postmortem_exported' | 'uptm.endpoint_added' | 'uptm.check_completed' | 'uptm.uptime_reported' | 'uptm.sla_exported' | 'smig.migration_applied' | 'smig.migration_rolled_back' | 'smig.diff_generated' | 'smig.schema_optimized' | 'qtun.query_analyzed' | 'qtun.index_suggested' | 'qtun.optimization_applied' | 'qtun.report_exported' | 'bsched.backup_completed' | 'bsched.restore_completed' | 'bsched.retention_enforced' | 'bsched.schedule_updated' | 'rplmgr.node_added' | 'rplmgr.failover_completed' | 'rplmgr.lag_detected' | 'rplmgr.sync_verified' | 'plmgr.pool_configured' | 'plmgr.connections_drained' | 'plmgr.stats_recorded' | 'plmgr.pool_optimized' | 'vscan.scan_completed' | 'vscan.vuln_found' | 'vscan.vuln_patched' | 'vscan.report_exported' | 'crot.credential_rotated' | 'crot.rotation_scheduled' | 'crot.rotation_failed' | 'crot.vault_synced' | 'caud.audit_completed' | 'caud.control_checked' | 'caud.report_generated' | 'caud.remediation_applied' | 'rbac.role_created' | 'rbac.assignment_granted' | 'rbac.assignment_revoked' | 'rbac.access_denied' | 'penf.policy_evaluated' | 'penf.decision_logged' | 'penf.policy_updated' | 'penf.violation_detected' | 'mrel.channel_created' | 'mrel.message_relayed' | 'mrel.dlq_processed' | 'mrel.batch_flushed' | 'sing.partition_assigned' | 'sing.checkpoint_saved' | 'sing.lag_detected' | 'sing.rebalanced' | 'ertr.rule_created' | 'ertr.event_routed' | 'ertr.fanout_completed' | 'ertr.dead_lettered' | 'qmgr.queue_created' | 'qmgr.message_dequeued' | 'qmgr.depth_exceeded' | 'qmgr.metrics_recorded' | 'psgw.topic_created' | 'psgw.subscription_added' | 'psgw.message_published' | 'psgw.ack_timeout' | 'cbld.image_built' | 'cbld.layers_optimized' | 'cbld.scan_completed' | 'cbld.config_exported' | 'ireg.repo_created' | 'ireg.image_pushed' | 'ireg.gc_completed' | 'ireg.tags_listed' | 'orch.deployed' | 'orch.scaled' | 'orch.rolled_back' | 'orch.health_checked' | 'smsh.route_created' | 'smsh.policy_applied' | 'smsh.breaker_toggled' | 'smsh.traffic_shifted' | 'cfmg.config_set' | 'cfmg.config_retrieved' | 'cfmg.config_rolled_back' | 'cfmg.config_exported' | 'sidx.index_created' | 'sidx.docs_indexed' | 'sidx.query_executed' | 'sidx.index_optimized' | 'anle.dataset_created' | 'anle.query_executed' | 'anle.data_ingested' | 'anle.cache_hit' | 'dlkh.table_created' | 'dlkh.snapshot_taken' | 'dlkh.compaction_completed' | 'dlkh.data_loaded' | 'etlp.job_started' | 'etlp.job_completed' | 'etlp.job_failed' | 'etlp.rows_processed' | 'rgen.report_generated' | 'rgen.template_created' | 'rgen.report_scheduled' | 'rgen.report_exported' | 'nrtr.route_created' | 'nrtr.policy_applied' | 'nrtr.traffic_routed' | 'nrtr.failover_triggered' | 'dngw.domain_resolved' | 'dngw.record_created' | 'dngw.cache_warmed' | 'dngw.query_completed' | 'lbor.backend_added' | 'lbor.health_checked' | 'lbor.traffic_distributed' | 'lbor.backend_drained' | 'cdnp.cache_populated' | 'cdnp.content_purged' | 'cdnp.origin_configured' | 'cdnp.edge_optimized' | 'rtcl.rule_created' | 'rtcl.client_limited' | 'rtcl.limit_updated' | 'rtcl.client_blocked'
+  | 'lgst.stream_created'
+  | 'lgst.alert_triggered'
+  | 'lgst.logs_exported'
+  | 'lgst.retention_rotated'
+  | 'mhub.metric_registered'
+  | 'mhub.alert_fired'
+  | 'mhub.metrics_exported'
+  | 'mhub.window_aggregated'
+  | 'evcr.pattern_detected'
+  | 'evcr.incident_created'
+  | 'evcr.root_cause_found'
+  | 'evcr.incident_resolved'
+  | 'trcl.trace_collected'
+  | 'trcl.bottleneck_found'
+  | 'trcl.service_mapped'
+  | 'trcl.traces_exported'
+  | 'dshb.panel_created'
+  | 'dshb.snapshot_taken'
+  | 'dshb.dashboard_shared'
+  | 'dshb.template_imported'
   | 'abtest.experiment_created'
   | 'abtest.variant_assigned'
   | 'abtest.conversion_recorded'
@@ -1479,6 +1504,11 @@ export function districtFor(kind: EidolonBuildingKind): District {
     case 'lb_orchestrator':
     case 'cdn_proxy':
     case 'rate_controller':
+    case 'log_streamer':
+    case 'metrics_hub':
+    case 'event_correlator':
+    case 'trace_collector':
+    case 'dashboard_builder':
       case 'network_firewall':
       case 'threat_detector':
       return 'industrial';